requirements.txt
numpy
opencv-python
insightface
onnxruntime
Pillow
torch